--- a/tests/unit/scripting.tcl
+++ b/tests/unit/scripting.tcl
@@ -147,7 +147,6 @@ start_server {tags {"scripting"}} {
     } {number 1}
 
     test {EVAL - Lua number -> Redis integer conversion} {
-        r set x 0
         run_script {
             local foo = redis.pcall('hincrby','hash','field',200000000)
             return {type(foo),foo}