@sundb 2024-03-06
if it is long long, what abotu create the string robj with long long, instead of converting to str and converting it to long long
in createStringObject() ?