@mdouglass 2024-03-06
> if it is long long, what abotu create the string robj with long long, instead of converting to str and converting it to long long
> in createStringObject() ?

Do you mean like this?
```
            if (double2ll((double)num, &lvalue)) {
                lua_argv[j] = createStringObjectFromLongLong(lvalue, lua_argv[j]);
                continue;
            } else {
```

Its a little bit awkward because of needing to skip the rest of the loop and I'm not clear on the consequences of skipping the code in the `/* Try to use a cached object. */` block since this is my first (well, now second) day with the redis codebase. Happy to do whichever version the team recommends.