@sundb 2024-03-07
> ```
>             if (double2ll((double)num, &lvalue)) {
>                 lua_argv[j] = createStringObjectFromLongLong(lvalue);
>                 continue;
>             } else {
> ```

i think it works.