@mdouglass 2024-03-07
> > ```
> >             if (double2ll((double)num, &lvalue)) {
> >                 lua_argv[j] = createStringObjectFromLongLong(lvalue);
> >                 continue;
> >             } else {
> > ```
> 
> i think it works, and please do a simple benchmark to see if it can bring benefit, thanks.

I don't know how to do benchmarks within redis and I can't find anything that documents it. Can you point me in the right direction for that?