@sundb 2024-03-07
> I don't know how to do benchmarks within redis and I can't find anything that documents it. Can you point me in the right direction for that?

Like this
```
redis-benchmark -n 100000 eval 'redis.call("del", "key"); for i = 1, 100 do redis.call("hincrby", "key", "field", 10000) end' 0
```